requirements.txt
streamlit
langgraph
langchain
langchain-core
langchain-groq
sentence-transformers
pypdf2
python-docx
chromadb
faiss-cpu
python-dotenv
pydantic>=2.0